refactor(pack): refactor slots task voip xdslaccess
