+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/task/task.html
