+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/voipEcoFax/voipEcoFax.controller.js
@@ -56,6 +56,3 @@ export default class PackVoipEcoFaxCtrl {
       });
   }
 }
-angular
-  .module('managerApp')
-  .controller('PackVoipEcoFaxCtrl', PackVoipEcoFaxCtrl);
